Fixing top layer's shape type

+++ correlation_layer.py
@@ -33,8 +33,8 @@ class CorrelationBaseOp(GpuOp, COp):
     paddedbottomheight = self.bottom_shape[2] + 2 * self.pad_size;
     paddedbottomwidth = self.bottom_shape[3] + 2 * self.pad_size;
 
-    self.top_width = np.ceil(float(paddedbottomwidth - border_size * 2) / float(self.stride1))
-    self.top_height = np.ceil(float(paddedbottomheight - border_size * 2) / float(self.stride1))
+    self.top_width = int(np.ceil(float(paddedbottomwidth - border_size * 2) / float(self.stride1)))
+    self.top_height = int(np.ceil(float(paddedbottomheight - border_size * 2) / float(self.stride1)))
 
     assert self.top_width >= 1
     assert self.top_height >= 1
